Refactored Bin
